:bug: The vscode multi dir projects cannot debug tests
A VSCode workspace with more than one directory cannot debug a test
using the command ceedlingExplorer.debugTestExecutable. A pop-up
appear containing the path to the debug executable without the last
part.

This commit fix the bug by using a global variable, common to every
CeedlingAdapter, to store the name of the currently debugged
executable.

fix #27