:bug: Look through all the adapters for a debug executable
The adapter factory is called on every workspace folders. Then there is
multiple instances of the adapter initiliazed when the debug function is
called. But only one of them has a debugTestExecutable set when the
debugTestExecutable command is called. We can use this to choose which
executable name to return.

This commit also fix the clean and clobber commands for workspaces with
multiple folders. Now the commands are executed on all the adapter
instances instead of the last initialized one.

fix #98