:sparkles: Use the ruby cli to evaluate the :build_root: path
Using ruby instead of a regex to get the env variable makes it more
robust and it should cover more usecases.

The feature is disabled by default because I'm not sure ruby is available in
all the setups and I don't want to break anything. It could also create some
security issues to execute ruby code automatically from vscode.